Moved methods placeShips() and checkShips()
